+++ awm/src/main/java/io/rightmesh/awm/collectors/InternetStatsCollector.java
@@ -1,8 +1,6 @@
-package io.rightmesh.awm;
+package io.rightmesh.awm.collectors;
 
 import android.content.Context;
-
-import io.rightmesh.awm.collectors.StatsCollector;
 
 public class InternetStatsCollector extends StatsCollector {
 
@@ -11,12 +9,12 @@ public class InternetStatsCollector extends StatsCollector {
     }
 
     @Override
-    void start() throws Exception {
+    public void start() throws Exception {
 
     }
 
     @Override
-    void stop() {
+    public void stop() {
 
     }
 }
