+++ awm/src/main/java/io/rightmesh/awm/collectors/StatsCollector.java
@@ -1,4 +1,4 @@
-package io.rightmesh.awm;
+package io.rightmesh.awm.collectors;
 
 import com.anadeainc.rxbus.Bus;
 import com.anadeainc.rxbus.BusProvider;
@@ -6,15 +6,17 @@ import com.vanniktech.rxpermission.Permission;
 
 import java.util.Map;
 
-abstract class StatsCollector {
+import io.rightmesh.awm.AwmComponent;
+
+public abstract class StatsCollector implements AwmComponent {
 
     Bus eventBus = BusProvider.getInstance();
     Map<String, Permission> permissions = null;
 
-    void setPermissions(Map<String, Permission> permissions) {
+    public void setPermissions(Map<String, Permission> permissions) {
         this.permissions = permissions;
     }
 
-    abstract void start() throws Exception;
-    abstract void stop();
+    public abstract void start() throws Exception;
+    public abstract void stop();
 }
