+++ awm/src/main/java/io/rightmesh/awm/collectors/WiFiAPStatsCollector.java
@@ -1,4 +1,4 @@
-package io.rightmesh.awm;
+package io.rightmesh.awm.collectors;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
@@ -10,7 +10,7 @@ import android.util.Log;
 
 import java.util.List;
 
-import io.rightmesh.awm.collectors.StatsCollector;
+import io.rightmesh.awm.stats.WiFiStats;
 
 public class WiFiAPStatsCollector extends StatsCollector {
 
@@ -21,7 +21,7 @@ public class WiFiAPStatsCollector extends StatsCollector {
     private WifiManager.WifiLock wifiLock;
     private WiFiScanReceiver wiFiScanReceiver;
 
-    WiFiAPStatsCollector(Context context) {
+    public WiFiAPStatsCollector(Context context) {
         this.context = context;
 
         wifiManager = (WifiManager) context.getApplicationContext()
@@ -33,7 +33,7 @@ public class WiFiAPStatsCollector extends StatsCollector {
     }
 
     @Override
-    void start() throws Exception {
+    public void start() throws Exception {
         if(wifiLock != null) {
             wifiLock.acquire();
         }
@@ -44,7 +44,7 @@ public class WiFiAPStatsCollector extends StatsCollector {
     }
 
     @Override
-    void stop() {
+    public void stop() {
         if(wifiLock != null) {
             wifiLock.release();
         }
