+++ awm/src/main/java/io/rightmesh/awm/stats/BluetoothStats.java
@@ -1,15 +1,17 @@
-package io.rightmesh.awm;
+package io.rightmesh.awm.stats;
 
 import android.bluetooth.BluetoothDevice;
 
 import java.io.Serializable;
 import java.util.Collection;
 
+import io.rightmesh.awm.stats.NetworkStat;
+
 public class BluetoothStats extends NetworkStat implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    BluetoothStats(Collection<BluetoothDevice> btDevices, String mac) {
+    public BluetoothStats(Collection<BluetoothDevice> btDevices, String mac) {
         this.mac = mac;
         for(BluetoothDevice device : btDevices) {
             macs.add(device.getAddress());
