+++ awm/src/main/java/io/rightmesh/awm/stats/NetworkStat.java
@@ -1,10 +1,11 @@
-package io.rightmesh.awm;
+package io.rightmesh.awm.stats;
 
 import java.util.HashSet;
 import java.util.Set;
 
 public abstract class NetworkStat {
 
+    String name;
     String mac;
     Set<String> macs = new HashSet<>();
 
@@ -15,6 +16,8 @@ public abstract class NetworkStat {
     public String getMac() {
         return mac;
     }
+
+    public String getName() { return name; }
 
     public Set<String> getMacs() {
         return macs;
