+++ awm/src/main/java/io/rightmesh/awm/stats/WiFiStats.java
@@ -1,12 +1,14 @@
-package io.rightmesh.awm;
+package io.rightmesh.awm.stats;
 
 import android.net.wifi.ScanResult;
 
 import java.util.Collection;
 
+import io.rightmesh.awm.stats.NetworkStat;
+
 public class WiFiStats extends NetworkStat {
 
-    WiFiStats(Collection<ScanResult> wifiDevices, String mac) {
+    public WiFiStats(Collection<ScanResult> wifiDevices, String mac) {
         this.mac = mac;
         for(ScanResult scan : wifiDevices) {
             macs.add(scan.BSSID);
